requirements.txt
anyio==3.6.2
aspose-cells==22.12.0
autopep8==2.0.0
Babel==2.11.0
boto3==1.26.26
botocore==1.29.26
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
defusedxml==0.7.1
et-xmlfile==1.1.0
fastapi==0.88.0
fonttools==4.38.0
google-auth==2.15.0
google-auth-oauthlib==0.7.1
gspread==5.7.2
h11==0.14.0
httplib2==0.21.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
joblib==1.2.0
JPype1==1.4.1
jsonpickle==3.0.0
numpy==1.23.5
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pdfminer.six==20221105
peewee==3.15.4
Pillow==9.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.10.2
pyparsing==3.0.9
pypdf2==2.11.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
uvloop==0.17.0
Wand==0.6.10
watchfiles==0.18.1
websockets==10.4
workbook==1.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
